+++ spec/bashly/commands/generate_spec.rb
@@ -229,11 +229,11 @@ describe Commands::Generate, :slow do
 
     let(:bashly_config_path) { "#{source_dir}/bashly.yml" }
     let(:bashly_config) { YAML.load_file bashly_config_path }
-    let(:watcher_double) { instance_double Filewatcher, watch: nil }
+    let(:watch_double) { instance_double Watch, on_change: nil }
 
     it 'generates immediately and on change' do
-      allow(Filewatcher).to receive(:new).and_return(watcher_double)
-      allow(watcher_double).to receive(:watch).and_yield
+      allow(Watch).to receive(:new).and_return(watch_double)
+      allow(watch_double).to receive(:on_change).and_yield
 
       expect { subject.execute %w[generate --watch] }
         .to output_approval('cli/generate/watch')
@@ -241,8 +241,8 @@ describe Commands::Generate, :slow do
 
     context 'when ConfigurationError is raised during watch' do
       it 'shows the error gracefully and continues to watch' do
-        allow(Filewatcher).to receive(:new).and_return(watcher_double)
-        allow(watcher_double).to receive(:watch) do |&block|
+        allow(Watch).to receive(:new).and_return(watch_double)
+        allow(watch_double).to receive(:on_change) do |&block|
           bashly_config['invalid_option'] = 'error this'
           File.write bashly_config_path, bashly_config.to_yaml
           block.call