+++ spec/bashly/integration/examples_spec.rb
@@ -5,41 +5,44 @@
 #
 # To only test examples containing a certain string in their path, run:
 # EXAMPLE=yaml bundle exec run spec examples
-
 describe 'generated bash scripts', :slow do
   # Make sure all examples are generated with strict mode
   before { ENV['BASHLY_STRICT'] = 'yes' }
   after { ENV['BASHLY_STRICT'] = nil }
 
   # Test public examples from the examples folder...
-  examples = Dir['examples/*'].select { |f| File.directory? f }
+  examples = Dir['examples/*'].select { |f| File.directory?(f) }
 
   # ...as well as internal examples, not suitable for public view
   fixtures = Dir['spec/fixtures/workspaces/*'].select do |f|
-    File.directory? f and File.exist? "#{f}/test.sh"
+    File.directory?(f) && File.exist?("#{f}/test.sh")
   end
 
   test_cases = fixtures + examples
 
   # Allow up to a certain string distance from the approval text in CI
-  leeway = ENV['CI'] ? 40 : 0
+  let(:leeway) { ENV['CI'] ? 40 : 0 }
 
   # For certain examples, allow some exceptions (replacements) since they
   # are too volatile (e.g. line number changes)
-  exceptions = {
-    'examples/stacktrace'    => [/download:\d+/, 'download:<line>'],
-    'examples/render-mandoc' => [/Version 0.1.0.*download\(1\)/, '<footer>'],
-  }
+  let(:exceptions) do
+    {
+      'examples/stacktrace'    => [/download:\d+/, 'download:<line>'],
+      'examples/render-mandoc' => [/Version 0.1.0.*download\(1\)/, '<footer>'],
+    }
+  end
 
   test_cases.each do |example|
-    approval_name = example.gsub 'spec/fixtures/workspaces', 'fixtures'
-
     next if ENV['EXAMPLE'] && !example.include?(ENV['EXAMPLE'])
 
     describe example do
+      let(:approval_name) do
+        example.gsub('spec/fixtures/workspaces', 'fixtures')
+      end
+
       it 'is executed properly' do
         output = 'not executed'
-        Dir.chdir example do
+        Dir.chdir(example) do
           output = `bash test.sh 2>&1`
         end